+++ packages/extension/src/topoEditor.ts
@@ -105,7 +105,11 @@ export class TopoEditorProvider implements vscode.CustomTextEditorProvider {
       else if (isEditMessage(message)) void controller.handleEdit(message);
       else if (isExportRequest(message)) void runExport(message.kind, document.uri);
       else if (isAgentGuideRequest(message)) {
-        void vscode.commands.executeCommand('topodraft.writeAgentGuide', message.saveAs === true);
+        void vscode.commands.executeCommand(
+          'topodraft.writeAgentGuide',
+          message.saveAs === true,
+          message.netbox === true,
+        );
       } else if (isNewFileRequest(message)) {
         void vscode.commands.executeCommand('topodraft.newFile', message.template);
       } else if (isListTemplatesRequest(message)) {